pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wallpaper-unpacker-gui"
version = "0.1.0"
description = "Offline Wallpaper Engine .pkg/.tex image extractor with GUI and CLI."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Wallpaper-Unpacker-GUI Contributors" }
]
dependencies = [
  "pillow",
  "lz4",
]

[project.scripts]
we-unpack = "main:main"

[tool.setuptools]
packages = ["app"]
py-modules = ["main", "gui"]
